BVN_Buoi5/backend/requirements.txt
flask
psycopg2-binary
faker
flask-cors